chore(vue): test file rename
